ClassLink Audio System/pc/ai_service/requirements.txt
google-generativeai>=0.3.0
SpeechRecognition>=3.10.0
gTTS>=2.4.0
PyPDF2>=3.0.0
python-docx>=1.1.0
pydub>=0.25.1
